+++ common/model/src/main/java/module-info.java
@@ -19,13 +19,14 @@
 /**
  * <p>Shared model classes.
  * <ul>
- * <li>{@link is.codion.common.model.UserPreferences}
  * <li>{@link is.codion.common.model.CancelException}
  * <li>{@link is.codion.common.model.filter.FilterModel}
  * <li>{@link is.codion.common.model.selection.SingleSelection}
  * <li>{@link is.codion.common.model.selection.MultiSelection}
  * <li>{@link is.codion.common.model.condition.ConditionModel}
  * <li>{@link is.codion.common.model.condition.TableConditionModel}
+ * <li>{@link is.codion.common.model.preferences.FilePreferencesFactory}
+ * <li>{@link is.codion.common.model.preferences.UserPreferences}
  * <li>{@link is.codion.common.model.summary.SummaryModel}
  * <li>{@link is.codion.common.model.summary.TableSummaryModel}
  * </ul>