+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityApplicationPanel.java
@@ -23,7 +23,7 @@ import is.codion.common.event.Event;
 import is.codion.common.i18n.Messages;
 import is.codion.common.logging.LoggerProxy;
 import is.codion.common.model.CancelException;
-import is.codion.common.model.UserPreferences;
+import is.codion.common.model.preferences.UserPreferences;
 import is.codion.common.observable.Observable;
 import is.codion.common.observable.Observer;
 import is.codion.common.property.PropertyStore;